+++ changes/202208111416.bugfix
@@ -1,1 +1,1 @@
-Refactor git for workspace service compatibility
+[git] modified interface to handle more generic references (tag, commit hash or branch name)
