+++ utils/hashing/hash_test.go
@@ -5,9 +5,11 @@
 package hashing
 
 import (
+	"fmt"
 	"strings"
 	"testing"
 
+	"github.com/bxcodec/faker/v3"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -47,5 +49,76 @@ func TestMd5(t *testing.T) {
 	}}
 	for _, testCase := range testCases {
 		assert.Equal(t, testCase.Hash, CalculateMD5Hash(testCase.Input))
+	}
+}
+
+func TestIsLikelyHexHashString(t *testing.T) {
+	tests := []struct {
+		input  string
+		isHash bool
+	}{
+		{
+			input:  "",
+			isHash: false,
+		},
+		{
+			input:  faker.Word(),
+			isHash: false,
+		},
+		{
+			input:  faker.Name(),
+			isHash: false,
+		},
+		{
+			input:  faker.Sentence(),
+			isHash: false,
+		},
+		{
+			input:  faker.CCNumber(),
+			isHash: false,
+		},
+		{
+			input:  faker.UUIDDigit(),
+			isHash: true,
+		},
+		{
+			input:  faker.UUIDHyphenated(),
+			isHash: false,
+		},
+		{
+			input:  faker.IPv4(),
+			isHash: false,
+		},
+		{
+			input:  faker.Paragraph(),
+			isHash: false,
+		},
+		{
+			input:  "1.0.1",
+			isHash: false,
+		},
+		{
+			input:  "v1.0.1",
+			isHash: false,
+		},
+		{
+			input:  CalculateMD5Hash(faker.Paragraph()),
+			isHash: true,
+		},
+
+		{
+			input:  CalculateHash(faker.Paragraph(), HashSha256),
+			isHash: true,
+		},
+		{
+			input:  "85817ddeed66c3e3805c73dbc7082de2674e349c",
+			isHash: true,
+		},
+	}
+	for i := range tests {
+		test := tests[i]
+		t.Run(fmt.Sprintf("%v_isHash(%v)", i, test.input), func(t *testing.T) {
+			require.Equal(t, test.isHash, IsLikelyHexHashString(test.input))
+		})
 	}
 }
